+++ issuer-application/src/main/java/ch/admin/bj/swiyu/issuer/infrastructure/web/management/CredentialController.java
@@ -7,6 +7,7 @@
 package ch.admin.bj.swiyu.issuer.infrastructure.web.management;
 
 import ch.admin.bj.swiyu.issuer.api.credentialoffer.CreateCredentialRequestDto;
+import ch.admin.bj.swiyu.issuer.api.credentialoffer.CredentialInfoResponseDto;
 import ch.admin.bj.swiyu.issuer.api.credentialoffer.CredentialWithDeeplinkResponseDto;
 import ch.admin.bj.swiyu.issuer.api.credentialofferstatus.StatusResponseDto;
 import ch.admin.bj.swiyu.issuer.api.credentialofferstatus.UpdateCredentialStatusRequestTypeDto;
@@ -17,7 +18,6 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -25,6 +25,7 @@ import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Map;
 import java.util.UUID;
 
 @RestController
@@ -73,52 +74,21 @@ public class CredentialController {
             responses = {
                     @ApiResponse(
                             responseCode = "200",
-                            description = "Credential offer found",
-                            content = @Content(
-                                    schema = @Schema(implementation = Object.class),
-                                    examples = {
-                                            @ExampleObject(
-                                                    name = "Offer data",
-                                                    summary = "Example of vc content/ offer data for vc with first/lastname as credentialSubjectData",
-                                                    value = """
-                                                            {"lastName":"Example","firstName":"Edward"}
-                                                            """
-                                            )
-                                    }
-                            )
+                            description = "Credential offer found"
                     )
             }
     )
-    public Object getCredentialOffer(@PathVariable UUID credentialId) {
-        return this.credentialService.getCredentialOffer(credentialId);
+    public CredentialInfoResponseDto getCredentialInformation(@PathVariable UUID credentialId) {
+        return this.credentialService.getCredentialOfferInformation(credentialId);
     }
 
+    @Deprecated(forRemoval = true)
     @Timed
+    /**
+     * Endpoint to retrieve the deeplink for a credential offer.
+     * @deprecated Use {@link #getCredentialInformation(UUID)} instead. Which contains the deeplink in the response.
+     */
     @GetMapping("/{credentialId}/offer_deeplink")
-    @Operation(
-            summary = "Get the offer deeplink",
-            responses = {
-                    @ApiResponse(
-                            responseCode = "200",
-                            description = "Offer deeplink found",
-                            content = @Content(
-                                    schema = @Schema(implementation = String.class),
-                                    examples = {
-                                            @ExampleObject(
-                                                    name = "Offer deeplink",
-                                                    summary = "Example of a deeplink",
-                                                    value = "swiyu://?credential_offer=%7B%22grants%22%3A%7B%22urn%3Aietf%3Aparams%3Aoauth%3Agrant-type%3Apre-authorized_code%22%3A%7B%22pre-authorized_code%22%3A%22b614c966-0c1d-4636-9aec-e2496d242d25%22%7D%7D%2C%22credential_issuer%22%3A%22https%3A%2F%2Fissuer-agent-oid4vci-d.bit.admin.ch%22%2C%22credential_configuration_ids%22%3A%5B%22myIssuerMetadataCredentialSupportedId%22%5D%7D"
-                                            )
-                                    }
-                            )
-                    ),
-                    @ApiResponse(
-                            responseCode = "404",
-                            description = "Offer not found or already expired",
-                            content = @Content(schema = @Schema(implementation = Object.class))
-                    )
-            }
-    )
     public String getCredentialOfferDeeplink(@PathVariable UUID credentialId) {
         return this.credentialService.getCredentialOfferDeeplink(credentialId);
     }
@@ -128,6 +98,27 @@ public class CredentialController {
     @Operation(summary = "Get the current status of an offer or the verifiable credential, if already issued.")
     public StatusResponseDto getCredentialStatus(@PathVariable UUID credentialId) {
         return this.credentialService.getCredentialStatus(credentialId);
+    }
+
+    @Timed
+    @PatchMapping("/{credentialId}")
+    @Operation(summary = "Update the status of an offer or the verifiable credential associated with the id. This is only for deferred flows. The status is set to ready for issuance",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "200",
+                            description = "Credential status updated",
+                            content = @Content(schema = @Schema(implementation = UpdateStatusResponseDto.class))
+                    ),
+                    @ApiResponse(
+                            responseCode = "400",
+                            description = "Bad request due to user content or internal call to external service like statuslist",
+                            content = @Content(schema = @Schema(implementation = Object.class))
+                    )
+            }
+    )
+    public UpdateStatusResponseDto updateCredentialForDeferredFlow(@PathVariable UUID credentialId, @RequestBody Map<String, Object> credentialOffer) {
+
+        return this.credentialService.updateOfferDataForDeferred(credentialId, credentialOffer);
     }
 
     @Timed