+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/domain/credentialoffer/ClientAgentInfo.java
@@ -1,0 +1,9 @@
+package ch.admin.bj.swiyu.issuer.domain.credentialoffer;
+
+public record ClientAgentInfo(
+        String remoteAddr,
+        String userAgent,
+        String acceptLanguage,
+        String acceptEncoding
+) {
+}