+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/domain/credentialoffer/CredentialOffer.java
@@ -116,6 +116,12 @@ public class CredentialOffer {
     private String holderJWK;
 
     /**
+     * Value used to store client agent infos for the deferred flow
+     */
+    @JdbcTypeCode(SqlTypes.JSON)
+    private ClientAgentInfo clientAgentInfo;
+
+    /**
      * Expiration in unix epoch (since 1.1.1970) timestamp in seconds
      */
     private long tokenExpirationTimestamp;
@@ -228,12 +234,21 @@ public class CredentialOffer {
 
     public void markAsDeferred(UUID transactionId,
                                CredentialRequestClass credentialRequest,
-                               String holderPublicKey) {
+                               String holderPublicKey,
+                               ClientAgentInfo clientAgentInfo) {
         this.credentialStatus = CredentialStatusType.DEFERRED;
         this.credentialRequest = credentialRequest;
         this.transactionId = transactionId;
         this.holderJWK = holderPublicKey;
+        this.clientAgentInfo = clientAgentInfo;
         log.info("Deferred credential response for offer {}. Management-ID is {} and status is {}. ",
+                this.metadataCredentialSupportedId, this.id, this.credentialStatus);
+    }
+
+    public void markAsReadyForIssuance(Map<String, Object> offerData) {
+        this.credentialStatus = CredentialStatusType.READY;
+        this.setOfferData(offerData);
+        log.info("Deferred Credential ready for issuance for offer {}. Management-ID is {} and status is {}. ",
                 this.metadataCredentialSupportedId, this.id, this.credentialStatus);
     }
 
@@ -253,5 +268,6 @@ public class CredentialOffer {
         this.transactionId = null;
         this.credentialRequest = null;
         this.holderJWK = null;
+        this.clientAgentInfo = null;
     }
 }