+++ issuer-service/src/test/java/ch/admin/bj/swiyu/issuer/oid4vci/service/CredentialServiceTest.java
@@ -7,18 +7,24 @@
 package ch.admin.bj.swiyu.issuer.oid4vci.service;
 
 import ch.admin.bj.swiyu.issuer.api.credentialoffer.CreateCredentialRequestDto;
+import ch.admin.bj.swiyu.issuer.api.credentialoffer.CredentialInfoResponseDto;
 import ch.admin.bj.swiyu.issuer.api.credentialofferstatus.CredentialStatusTypeDto;
 import ch.admin.bj.swiyu.issuer.api.credentialofferstatus.UpdateCredentialStatusRequestTypeDto;
+import ch.admin.bj.swiyu.issuer.api.credentialofferstatus.UpdateStatusResponseDto;
 import ch.admin.bj.swiyu.issuer.api.oid4vci.CredentialRequestDto;
+import ch.admin.bj.swiyu.issuer.api.oid4vci.DeferredCredentialRequestDto;
 import ch.admin.bj.swiyu.issuer.common.config.ApplicationProperties;
 import ch.admin.bj.swiyu.issuer.common.config.OpenIdIssuerConfiguration;
 import ch.admin.bj.swiyu.issuer.common.exception.BadRequestException;
 import ch.admin.bj.swiyu.issuer.common.exception.OAuthException;
+import ch.admin.bj.swiyu.issuer.common.exception.Oid4vcException;
 import ch.admin.bj.swiyu.issuer.domain.credentialoffer.*;
+import ch.admin.bj.swiyu.issuer.domain.openid.credentialrequest.CredentialRequestClass;
 import ch.admin.bj.swiyu.issuer.domain.openid.metadata.CredentialClaim;
 import ch.admin.bj.swiyu.issuer.domain.openid.metadata.CredentialConfiguration;
 import ch.admin.bj.swiyu.issuer.domain.openid.metadata.IssuerMetadataTechnical;
 import ch.admin.bj.swiyu.issuer.service.*;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
@@ -37,11 +43,11 @@ import static java.time.Instant.now;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 class CredentialServiceTest {
     private final Map<String, Object> offerData = Map.of("hello", "world");
+    private final ObjectMapper objectMapper = new ObjectMapper();
     @Mock
     CredentialOfferRepository credentialOfferRepository;
     CredentialService credentialService;
@@ -58,6 +64,7 @@ class CredentialServiceTest {
     private CredentialOffer suspended;
     private StatusList statusList;
     private CreateCredentialRequestDto createCredentialRequestDto;
+    private CredentialFormatFactory credentialFormatFactory;
 
     @BeforeEach
     void setUp() {
@@ -65,7 +72,7 @@ class CredentialServiceTest {
         statusListService = Mockito.mock(StatusListService.class);
         KeyAttestationService keyAttestationService = Mockito.mock(KeyAttestationService.class);
         issuerMetadata = Mockito.mock(IssuerMetadataTechnical.class);
-        CredentialFormatFactory credentialFormatFactory = Mockito.mock(CredentialFormatFactory.class);
+        credentialFormatFactory = Mockito.mock(CredentialFormatFactory.class);
         applicationProperties = Mockito.mock(ApplicationProperties.class);
         OpenIdIssuerConfiguration openIdIssuerConfiguration = Mockito.mock(OpenIdIssuerConfiguration.class);
         dataIntegrityService = Mockito.mock(DataIntegrityService.class);
@@ -143,13 +150,14 @@ class CredentialServiceTest {
 
         // note: getting an expired offer will immediately update it to expired
         var expiredOfferId = expiredOffer.getId();
-        var response = credentialService.getCredentialOffer(expiredOfferId);
+        var response = credentialService.getCredentialOfferInformation(expiredOfferId);
 
         Mockito.verify(credentialOfferRepository, Mockito.times(1)).findByIdForUpdate(expiredOfferId);
         Mockito.verify(credentialOfferRepository, Mockito.times(1)).save(any());
 
         // offer data should be null after expiration therefore no offer data or deeplink should be returned
-        assertNull(response);
+        assertNull(response.holderJWK());
+        assertNull(response.clientAgentInfo());
 
         var statusResponse = credentialService.getCredentialStatus(expiredOfferId);
 
@@ -170,10 +178,11 @@ class CredentialServiceTest {
     @Test
     void getCredentialOfferWhenNotExpired_thenSuccess() {
 
-        Map<String, String> response = (Map<String, String>) credentialService.getCredentialOffer(valid.getId());
+        when(applicationProperties.getDeeplinkSchema()).thenReturn("test-swiyu");
+        CredentialInfoResponseDto response = credentialService.getCredentialOfferInformation(valid.getId());
 
         assertNotNull(response);
-        assertEquals(offerData.get("hello"), response.get("hello"));
+        assertTrue(response.offerDeeplink().startsWith("test-swiyu://"));
 
         credentialService.getCredentialOfferDeeplink(valid.getId());
     }
@@ -185,7 +194,7 @@ class CredentialServiceTest {
 
         when(credentialOfferStatusRepository.findByOfferStatusId(issued.getId())).thenReturn(offerStatusSet);
 
-        Mockito.doNothing().when(statusListService).revoke(offerStatusSet);
+        doNothing().when(statusListService).revoke(offerStatusSet);
 
         var updated = credentialService.updateCredentialStatus(issued.getId(), UpdateCredentialStatusRequestTypeDto.REVOKED);
 
@@ -222,20 +231,15 @@ class CredentialServiceTest {
         var uuid = UUID.randomUUID();
 
         var expirationTimeStamp = now().plusSeconds(1000).getEpochSecond();
-        var offer = getCredentialOffer(CredentialStatusType.IN_PROGRESS, expirationTimeStamp, offerData, uuid, uuid, UUID.randomUUID());
+        var offer = getCredentialOffer(CredentialStatusType.IN_PROGRESS, expirationTimeStamp, offerData, uuid, uuid, UUID.randomUUID(), Map.of(), null);
         offer.setTokenExpirationTimestamp(Instant.now().minusSeconds(600).getEpochSecond());
 
         when(credentialOfferRepository.findByAccessToken(uuid)).thenReturn(Optional.of(offer));
 
         // WHEN credential is created for offer with expired timestamp
-        var credentialRequestDto = new CredentialRequestDto(
-                "vc+sd-jwt",
-                new HashMap<>(),
-                null
-        );
-        var uuidString = uuid.toString();
-        var ex = assertThrows(OAuthException.class, () ->
-                credentialService.createCredential(credentialRequestDto, uuidString));
+        var credentialRequestDto = getCredentialRequestDto();
+        var accessToken = uuid.toString();
+        var ex = assertThrows(OAuthException.class, () -> credentialService.createCredential(credentialRequestDto, accessToken, null));
 
         // THEN Status is changed and offer data is cleared
         assertEquals("INVALID_REQUEST", ex.getError().toString());
@@ -249,19 +253,15 @@ class CredentialServiceTest {
         var preAuthorizedCode = UUID.randomUUID();
 
         var expirationTimeStamp = Instant.now().minusSeconds(10).getEpochSecond();
-        var offer = getCredentialOffer(CredentialStatusType.OFFERED, expirationTimeStamp, offerData, uuid, preAuthorizedCode, UUID.randomUUID());
+        var offer = getCredentialOffer(CredentialStatusType.OFFERED, expirationTimeStamp, offerData, uuid, preAuthorizedCode, UUID.randomUUID(), Map.of(), null);
 
         when(credentialOfferRepository.findByAccessToken(uuid)).thenReturn(Optional.of(offer));
 
         // WHEN credential is created for offer with expired timestamp
-        var credentialRequestDto = new CredentialRequestDto(
-                "vc+sd-jwt",
-                new HashMap<>(),
-                null
-        );
+        var credentialRequestDto = getCredentialRequestDto();
         var uuidString = uuid.toString();
         var ex = assertThrows(OAuthException.class, () ->
-                credentialService.createCredential(credentialRequestDto, uuidString));
+                credentialService.createCredential(credentialRequestDto, uuidString, null));
 
         // THEN Status is changed and offer data is cleared
         assertEquals(CredentialStatusType.EXPIRED, offer.getCredentialStatus());
@@ -274,7 +274,7 @@ class CredentialServiceTest {
     void givenExpiredOffer_whenTokenIsCreated_throws() {
         var uuid = UUID.randomUUID();
         var expirationTimeStamp = Instant.now().minusSeconds(10).getEpochSecond();
-        var offer = getCredentialOffer(CredentialStatusType.OFFERED, expirationTimeStamp, offerData, uuid, uuid, UUID.randomUUID());
+        var offer = getCredentialOffer(CredentialStatusType.OFFERED, expirationTimeStamp, offerData, uuid, uuid, UUID.randomUUID(), Map.of(), null);
 
         when(credentialOfferRepository.findByPreAuthorizedCode(uuid)).thenReturn(Optional.of(offer));
 
@@ -297,7 +297,7 @@ class CredentialServiceTest {
 
         when(credentialOfferStatusRepository.findByOfferStatusId(issued.getId())).thenReturn(offerStatusSet);
 
-        Mockito.doNothing().when(statusListService).revoke(offerStatusSet);
+        doNothing().when(statusListService).revoke(offerStatusSet);
 
         credentialService.updateCredentialStatus(issued.getId(), UpdateCredentialStatusRequestTypeDto.REVOKED);
 
@@ -311,7 +311,7 @@ class CredentialServiceTest {
 
         when(credentialOfferStatusRepository.findByOfferStatusId(issued.getId())).thenReturn(offerStatusSet);
 
-        Mockito.doNothing().when(statusListService).revoke(offerStatusSet);
+        doNothing().when(statusListService).revoke(offerStatusSet);
 
         credentialService.updateCredentialStatus(issued.getId(), UpdateCredentialStatusRequestTypeDto.SUSPENDED);
 
@@ -325,7 +325,7 @@ class CredentialServiceTest {
 
         when(credentialOfferStatusRepository.findByOfferStatusId(suspended.getId())).thenReturn(offerStatusSet);
 
-        Mockito.doNothing().when(statusListService).revoke(offerStatusSet);
+        doNothing().when(statusListService).revoke(offerStatusSet);
 
         credentialService.updateCredentialStatus(suspended.getId(), UpdateCredentialStatusRequestTypeDto.ISSUED);
 
@@ -369,6 +369,44 @@ class CredentialServiceTest {
                 credentialService.createCredential(createCredentialRequestDto));
 
         assertTrue(exception.getMessage().contains("Credential offer metadata test-metadata is not supported - should be one of different-test-metadata"));
+    }
+
+    @Test
+    void testCreateCredentialOffer_invalidFormat_thenBadRequest() {
+
+        var credentialRequestDto = getCredentialRequestDto();
+        var clientInfo = getClientInfo();
+        var claim = new CredentialClaim();
+        claim.setMandatory(true);
+        claim.setValueType("string");
+        var credConfig = mock(CredentialConfiguration.class);
+        when(credConfig.getCredentialDefinition()).thenReturn(null);
+        when(credConfig.getClaims()).thenReturn(Map.of("hello", claim));
+        when(credConfig.getFormat()).thenReturn("not-vc+sd-jwt");
+        when(credConfig.getVct()).thenReturn("test-vct");
+
+        var credentialOffer = getCredentialOffer(
+                CredentialStatusType.IN_PROGRESS,
+                Instant.now().plusSeconds(600).getEpochSecond(),
+                offerData,
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                Map.of("deferred", false),
+                null);
+
+        when(statusListService.findByUriIn(any())).thenReturn(List.of(statusList));
+        when(credentialOfferStatusRepository.save(any())).thenReturn(getCredentialOfferStatus(UUID.randomUUID(), UUID.randomUUID()));
+        when(credentialOfferRepository.findByIdForUpdate(any(UUID.class))).thenReturn(Optional.empty());
+        when(issuerMetadata.getCredentialConfigurationSupported()).thenReturn(Map.of("different-test-metadata", mock(CredentialConfiguration.class)));
+        when(credentialOfferRepository.findByAccessToken(credentialOffer.getAccessToken())).thenReturn(Optional.of(credentialOffer));
+        when(issuerMetadata.getCredentialConfigurationById("test")).thenReturn(credConfig);
+
+        var accessToken = credentialOffer.getAccessToken().toString();
+        var exception = assertThrows(Oid4vcException.class, () ->
+                credentialService.createCredential(credentialRequestDto, accessToken, clientInfo));
+
+        assertTrue(exception.getMessage().contains("Mismatch between requested and offered format."));
     }
 
     @Test
@@ -550,7 +588,7 @@ class CredentialServiceTest {
         when(credConfig.getCryptographicBindingMethodsSupported()).thenReturn(List.of("did:jwk", "jwk"));
         when(issuerMetadata.getCredentialConfigurationSupported()).thenReturn(Map.of("test-metadata", credConfig));
         when(issuerMetadata.getCredentialConfigurationById("test-metadata")).thenReturn(credConfig);
-        when(dataIntegrityService.getVerifiedOfferData(any(), any())).thenReturn(Map.of("hello", "world"));
+        when(dataIntegrityService.getVerifiedOfferData(any(), any())).thenReturn(offerData);
         when(credentialOfferRepository.save(any())).thenAnswer(new Answer<CredentialOffer>() {
             @Override
             public CredentialOffer answer(InvocationOnMock invocation) {
@@ -559,6 +597,212 @@ class CredentialServiceTest {
         });
 
         assertDoesNotThrow(() -> credentialService.createCredential(createCredentialRequestDto));
+    }
+
+    @Test
+    void testCreateCredential_deferred() throws JsonProcessingException {
+
+        var credentialRequestDto = getCredentialRequestDto();
+        var clientInfo = getClientInfo();
+
+        CredentialOffer credentialOffer = getCredentialOffer(
+                CredentialStatusType.IN_PROGRESS,
+                Instant.now().plusSeconds(600).getEpochSecond(),
+                offerData,
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                Map.of("deferred", true),
+                null);
+
+        when(credentialOfferRepository.findByAccessToken(credentialOffer.getAccessToken())).thenReturn(Optional.of(credentialOffer));
+        when(credentialOfferRepository.findByPreAuthorizedCode(any(UUID.class))).thenReturn(Optional.empty());
+
+        // Mock the factory to return the builder
+        var sdJwtCredential = mock(SdJwtCredential.class);
+        when(credentialFormatFactory.getFormatBuilder(anyString())).thenReturn(sdJwtCredential);
+        when(sdJwtCredential.credentialOffer(any())).thenReturn(sdJwtCredential);
+        when(sdJwtCredential.credentialResponseEncryption(any())).thenReturn(sdJwtCredential);
+        when(sdJwtCredential.holderBinding(any())).thenReturn(sdJwtCredential);
+        when(sdJwtCredential.credentialType(any())).thenReturn(sdJwtCredential);
+        when(statusListService.findByUriIn(any())).thenReturn(List.of(statusList));
+
+        var claim = new CredentialClaim();
+        claim.setMandatory(true);
+        claim.setValueType("string");
+        var credConfig = mock(CredentialConfiguration.class);
+        when(credConfig.getCredentialDefinition()).thenReturn(null);
+        when(credConfig.getClaims()).thenReturn(Map.of("hello", claim));
+        when(credConfig.getFormat()).thenReturn("vc+sd-jwt");
+        when(credConfig.getVct()).thenReturn("test-vct");
+
+        when(credConfig.getCryptographicBindingMethodsSupported()).thenReturn(List.of("did:jwk", "jwk"));
+        when(issuerMetadata.getCredentialConfigurationSupported()).thenReturn(Map.of("test", credConfig));
+        when(issuerMetadata.getCredentialConfigurationById(any())).thenReturn(credConfig);
+
+        credentialService.createCredential(credentialRequestDto, credentialOffer.getAccessToken().toString(), clientInfo);
+
+        // check if is issued && data removed
+        assertEquals(CredentialStatusType.DEFERRED, credentialOffer.getCredentialStatus());
+        String clientInfoString = objectMapper.writeValueAsString(clientInfo);
+        verify(webhookService).produceDeferredEvent(credentialOffer.getId(), clientInfoString);
+    }
+
+    @Test
+    void updateOfferDataForDeferred_shouldUpdateOfferData_whenDeferredAndInDeferredState() {
+        // Arrange
+        UUID credentialId = UUID.randomUUID();
+        CredentialOffer credentialOffer = mock(CredentialOffer.class);
+
+        when(credentialOffer.isDeferred()).thenReturn(true);
+        when(credentialOffer.getCredentialStatus()).thenReturn(CredentialStatusType.DEFERRED);
+        when(credentialOfferRepository.findByIdForUpdate(credentialId)).thenReturn(Optional.of(credentialOffer));
+        doNothing().when(credentialOffer).markAsReadyForIssuance(any());
+        when(credentialOfferRepository.save(credentialOffer)).thenReturn(credentialOffer);
+
+        UpdateStatusResponseDto response = credentialService.updateOfferDataForDeferred(credentialId, offerData);
+
+        verify(credentialOfferRepository).findByIdForUpdate(credentialId);
+        verify(credentialOffer).markAsReadyForIssuance(any());
+        verify(credentialOfferRepository).save(credentialOffer);
+        assertNotNull(response);
+    }
+
+    @Test
+    void updateOfferDataForDeferred_shouldThrow_whenNotDeferredOrNotInDeferredState() {
+        // Arrange
+        UUID credentialId = UUID.randomUUID();
+        Map<String, Object> offerDataMap = Map.of("claim", "value");
+        CredentialOffer credentialOffer = mock(CredentialOffer.class);
+
+        when(credentialOffer.isDeferred()).thenReturn(false);
+        when(credentialOffer.getCredentialStatus()).thenReturn(CredentialStatusType.DEFERRED);
+        when(credentialOfferRepository.findByIdForUpdate(credentialId)).thenReturn(Optional.of(credentialOffer));
+
+        // Act & Assert
+        assertThrows(BadRequestException.class, () -> credentialService.updateOfferDataForDeferred(credentialId, offerDataMap));
+    }
+
+    @Test
+    void updateOfferDataForDeferred_shouldUpdateCredentialOfferData() {
+        // Arrange
+        UUID credentialId = UUID.randomUUID();
+        Map<String, Object> offerDataMap = Map.of("claim", "value");
+        CredentialOffer credentialOffer = getCredentialOffer(
+                CredentialStatusType.DEFERRED,
+                Instant.now().plusSeconds(600).getEpochSecond(),
+                Map.of(),
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                Map.of("deferred", true),
+                null);
+
+        when(credentialOfferRepository.findByIdForUpdate(credentialId)).thenReturn(Optional.of(credentialOffer));
+        when(credentialOfferRepository.save(credentialOffer)).thenReturn(credentialOffer);
+
+        credentialService.updateOfferDataForDeferred(credentialId, offerDataMap);
+
+        verify(credentialOfferRepository, times(1)).save(credentialOffer);
+    }
+
+    @Test
+    void testCreateCredentialFromDeferredRequest_notReady_thenException() {
+
+        UUID accessToken = UUID.randomUUID();
+
+        CredentialOffer credentialOffer = getCredentialOffer(
+                CredentialStatusType.DEFERRED,
+                Instant.now().plusSeconds(600).getEpochSecond(),
+                Map.of(),
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                Map.of("deferred", true),
+                UUID.randomUUID());
+
+        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(credentialOffer.getTransactionId(), new HashMap<>());
+
+        when(credentialOfferRepository.findByAccessToken(accessToken)).thenReturn(Optional.of(credentialOffer));
+        when(credentialOfferRepository.findByPreAuthorizedCode(any(UUID.class))).thenReturn(Optional.empty());
+
+        // Act
+        var accessTokenString = accessToken.toString();
+        var exception = assertThrows(Oid4vcException.class, () ->
+                credentialService.createCredentialFromDeferredRequest(deferredRequest, accessTokenString));
+
+        assertEquals("The credential is not marked as ready to be issued", exception.getMessage());
+    }
+
+    @Test
+    void testCreateCredentialFromDeferredRequest_accesTokenExpired_thenException() {
+
+        UUID transactionId = UUID.randomUUID();
+        UUID accessToken = UUID.randomUUID();
+
+        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(transactionId, new HashMap<>());
+
+        CredentialOffer credentialOffer = getCredentialOffer(
+                CredentialStatusType.DEFERRED,
+                Instant.now().minusSeconds(600).getEpochSecond(),
+                Map.of(),
+                accessToken,
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                Map.of("deferred", true),
+                transactionId);
+
+        when(credentialOfferRepository.findByAccessToken(accessToken)).thenReturn(Optional.of(credentialOffer));
+
+        // Act
+        var accessTokenString = accessToken.toString();
+        var exception = assertThrows(OAuthException.class, () ->
+                credentialService.createCredentialFromDeferredRequest(deferredRequest, accessTokenString));
+
+        assertEquals("Invalid accessToken", exception.getMessage());
+    }
+
+    @Test
+    void testCreateCredentialFromDeferredRequest_success() {
+
+        UUID transactionId = UUID.randomUUID();
+        UUID accessToken = UUID.randomUUID();
+
+        DeferredCredentialRequestDto deferredRequest = new DeferredCredentialRequestDto(transactionId, new HashMap<>());
+
+        CredentialOffer credentialOffer = getCredentialOffer(
+                CredentialStatusType.READY,
+                Instant.now().plusSeconds(600).getEpochSecond(),
+                offerData,
+                accessToken,
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                Map.of("deferred", true),
+                transactionId);
+
+        when(credentialOfferRepository.findByAccessToken(accessToken)).thenReturn(Optional.of(credentialOffer));
+        when(credentialOfferRepository.findByPreAuthorizedCode(any(UUID.class))).thenReturn(Optional.empty());
+
+        // Mock the factory to return the builder
+        var sdJwtCredential = mock(SdJwtCredential.class);
+        when(credentialFormatFactory.getFormatBuilder(anyString())).thenReturn(sdJwtCredential);
+        when(sdJwtCredential.credentialOffer(any())).thenReturn(sdJwtCredential);
+        when(sdJwtCredential.credentialResponseEncryption(any())).thenReturn(sdJwtCredential);
+        when(sdJwtCredential.holderBinding(any())).thenReturn(sdJwtCredential);
+        when(sdJwtCredential.credentialType(any())).thenReturn(sdJwtCredential);
+
+        // Act
+        credentialService.createCredentialFromDeferredRequest(deferredRequest, accessToken.toString());
+
+        // check if is issued && data removed
+        assertEquals(CredentialStatusType.ISSUED, credentialOffer.getCredentialStatus());
+        assertNull(credentialOffer.getOfferData());
+        assertNull(credentialOffer.getTransactionId());
+        assertNull(credentialOffer.getHolderJWK());
+        assertNull(credentialOffer.getClientAgentInfo());
+
+        verify(credentialOfferRepository).save(credentialOffer);
+        verify(webhookService).produceStateChangeEvent(any(), any());
     }
 
     private @NotNull Set<CredentialOfferStatus> getCredentialOfferStatusSet() {
@@ -574,18 +818,19 @@ class CredentialServiceTest {
 
     private CredentialOffer getCredentialOffer(CredentialStatusType statusType, long offerExpirationTimestamp, Map<String, Object> offerData) {
 
-        return getCredentialOffer(statusType, offerExpirationTimestamp, offerData, UUID.randomUUID(), UUID.randomUUID(), UUID.randomUUID());
+        return getCredentialOffer(statusType, offerExpirationTimestamp, offerData, UUID.randomUUID(), UUID.randomUUID(), UUID.randomUUID(), Map.of(), null);
     }
 
-    private CredentialOffer getCredentialOffer(CredentialStatusType status, long offerExpirationTimestamp, Map<String, Object> offerData, UUID accessToken, UUID preAuthorizedCode, UUID nonce) {
+    private CredentialOffer getCredentialOffer(CredentialStatusType status, long offerExpirationTimestamp, Map<String, Object> offerData, UUID accessToken, UUID preAuthorizedCode, UUID nonce, Map<String, Object> offerMetadata, UUID transactionId) {
 
         return new CredentialOffer(
                 UUID.randomUUID(),
                 status,
-                Collections.emptyList(),
+                List.of("test"),
                 offerData,
-                new HashMap<>(),
+                offerMetadata,
                 accessToken,
+                transactionId,
                 null,
                 null,
                 Instant.now().plusSeconds(600).getEpochSecond(),
@@ -594,7 +839,19 @@ class CredentialServiceTest {
                 offerExpirationTimestamp,
                 Instant.now(),
                 Instant.now(),
+                new CredentialRequestClass("vc+sd-jwt", null, null)
+        );
+    }
+
+    private @NotNull CredentialRequestDto getCredentialRequestDto() {
+        return new CredentialRequestDto(
+                "vc+sd-jwt",
+                new HashMap<>(),
                 null
         );
+    }
+
+    private @NotNull ClientAgentInfo getClientInfo() {
+        return new ClientAgentInfo("test-agent", "1.0", "test-client", "test-client-id");
     }
 }